Add a basic 'experiments' app for recording the results of experiments
The motivation for this is that we need to record more information
than can be easily recorded in Google Analytics alone for the MIT
county performance page experiment.  For example, they want to be able
to cross-reference particular survey responses with any actions on the
page earlier, so we need to record some unique ID for them.

However, this might well be useful as a framework for other experiemnts
in the future in any case.