MIT: record passed-in parameters and actions taken
We will be supplied the following parameters from the Facebook
advertisements:

  g: Gender, either 'm' or 'f'
  agroup: Age group, either 'u' (under) or 'o' (over) some threshold

There will also be a 'variant' parameter indicating which variant of
the page is being used.  Finally, we create a "user key", which is really
just a random number associated with the original page view).

All of these parameters need to be recorded as events in the experiments
app whenever the user submits one fo the forms on that page.  Some of
this data (user_key, variant) first into columns in the
experiments_events table anyway, while the rest (g, agroup) are recorded
as JSON in its extra_data column.

As well as introducing those changes, this commit does some refactoring
to remove repetition in the form_valid method.