Add a JSON extra_data field to each experiment event
It would be nice to migrate to using django-hstore for this in the
future, but that would mean making sure that the Postgres hstore
extension is loaded, which involves some thinking + changes that I don't
have time for at the moment.  So, in the interim, just make sure that
people don't use this field for non-JSON data via the admin interface.