MIT: keep experiment data in the session rather than query parameters
Generating a user key on each view of the county performance page, and
passing it on as a parameter, means that if the same user returns to
the page later (e.g. by pressing "Back" after a sentate comment
submission) they'll probably have a different user key - this is
problematic, and we can easily get around it by storing the user key,
variant and facebook demographics in the session.