.travis.yml: avoid dependency problems that broke PostGIS template creation
Thanks to Matthew Somerville for doing much of the investigation needed
for this problem.  The use of the pgdg and ubuntugis PPAs created an
unresolvable set of dependencies if you wanted both the python-gdal and
a working postgresql-9.1-postgis-* to be installed.  The simplest
solution is to just remove these PPAs before installation, since the
tests all work fine with PostGIS 1.5.3 (the version in base Ubuntu
12.04).

Fixes #1448