ZA: Correct party percentages by excluding older members
The party counts in the National Assembly (say) that were being used to
calculate the proportion of the counts were including older member
(those from before the elections).  This commit changes this so that we
only count people associated with positions that are current.

This is a fix for the first part of #1450 - thanks to Geoff Kilpin for
spotting that problem.