Add a 'hidden' field to the Person model
This field has been introduced to reduce admin burden for Mzalendo:
county representatives, for example, aren't a focus of the site, but
because they have public pages on the site, they very frequently
request updates or corrections.

We want to keep these representatives in the database, since it's useful
data that they were aspirant or elected representatives, particurly if
they then go on to be an MP, for example.  So, a reasonable solution
seems to be to make it possible to mark certain people as "hidden" so
that their public person pages no longer are shown.