Use maybehidden to suppress generation of links to hidden people
Having lots of broken links on the site would be bad for our ranking on
Google and confusing to users, so this suppresses any links to people
who they wouldn't be able to see the people pages for.

I haven't added these to the 'featured people' pages since presumably
you wouldn't feature someone who is hidden.

Note that the number of templates that had to be changed does suggest
that there's a lot of redundancy in how people are displayed in
templates, which would be worth looking at in the future.