Add a script to restart constituency contact positions
We need this because it will be some time before we have new
constituency contact information.  It's reasonable to guess, however,
that people who were previously MPs or MPLs and constituency contacts
who have been re-elected will still be constituency contacts.  This
script finds those people and restarts their constituency contact
positions.

This is part of the fix for #1466.