Avoid running all core, hansard and search tests twice
Before switching to nose as our test runner, all the tests had to be
in a single 'tests' module in each application.  However, when using
nose it will find tests whereever they are and it'll find them twice
if you import them into __init__.py as well.  This slows down the
tests and produces double the output when there are errors, so leave
nose to find them via each module in the tests package.