MIT: add variants with social context to the County Performance pages
Our MIT collaborators have requested that we have add an additional 3
variants, which are identical to the existing three variants but with
some social context - an indication of how many other people have taken
part:

  "We would also like to add 3 new pages (i.e. we will be randomly
   assigning visitors from FB to one of 6 pages). These pages will be
   exact copies of the control, opportunity and threat pages from above
   and will ALL have the additional sentence (on it's own line
   immediately after the opportunity/threat paragraph, or by itself in
   the case of the control page).

   *342 people joined us in just 3 days!  Take action, as more and more
   citizens hold their county governments accountable.*"

This commit introduces this change.  Note that a new Google Analytics
experiment will need to be set up to switch between all 6 variants
rather than the original 3, and the GA experiment key set in the
COUNTY_PERFORMANCE_EXPERIMENT_KEY setting.