Make two URLs for the two County Performance experiments
We don't want to risk people following links from the earlier trial
experiment while the larger experiment is running and the data being
contaminated by that. The safest way to do that is to use two different
base URLs for the two experiments - one is now /county-performance and
the other is /county-performance-2

The commit introduces that change, while reusing the same view classes
with minor modifications.