Add a test suite to make sure we properly count people related to a place.
nb This test matches the current intent: as the code is written at the
moment, the count of people associated with the place contains every
person with a current position associated with this Place. This might
look confusing, since some of these people won't be findable through this
page, but that's probably another bug (either to change that displayed
count, or to include links to "other associated people").