Completely hide hidden people (don't even list them)
Previously we would still list people who were marked with hidden=True
and not link to their person page (which would 404 except for admins)
but this produced very unweildy lists of people on organisation or place
pages.

This commit changes this behaviour to not even list hidden people on
organisation or place pages.