Allow compass to be used via wsgi with DEBUG = True
If DEBUG is True, then django-pipeline will use compass to compile .scss
files on the fly, but for this to work we need compass to be on the PATH
and for GEM_HOME to correctly set in the environment.  This commit adds
a config option GEMS_DIRECTORY (which correponds to GEM_HOME) - if that
is set, then wsgi.py will set up those environment variables.