ZA: redirect person/../appearances/ to person/../
The person/../appearances/ url current displays 'No appearances were
found'. The correct fix is to redirect to the appearances tab on the
relevant person page, but this relies on issue #930 (ZA: person page
tab selection should change url). As a temporary fix this redirects the
user to the person page and avoids the error.