Suppress the warnings about missing loggers for elasticsearch
If there are no loggers defined for the elasticsearch Python package you
get errors like the following:

  No handlers could be found for logger "elasticsearch" (in production)
  No handlers could be found for logger "elasticsearch.trace" (local dev)

(I'm not sure why there's that difference in those logger names.)

This commit sets up loggers for both names in the Django settings module
that will send any logging of level INFO and above to standard error.
This suppresses that warning and should mean that we're able to see
real warnings and errors in the logs.