Add fields to Bill for linking to act PDF
Bills become acts, and when they do we want to link to the act
document instead of the bill. This adds two fields to the Bill
model that hold the act title and URL.
Additionally, we display the act title and link to the act document
if the fields are populated, instead of always showing the bill title.