Upgrade modernizr to a version supporting IE 11
The previous version of modernizr that we were using didn't support IE
11, which mean the carousel on the homepage was broken in that version.

Luckily modernizr builds have a "Build:" comment at the top which links
to the page using the same build options, hooray! The only change is the
'-iepp' option has been removed, but from what I've read the '-shiv'
option (which I've added to the new build) roughly replaces it, but
without print support.

I've also taken this opportunity to switch to an uncompressed version of
modernizr (though annoyingly it the download still has a compressed
version of yepnope at the end of the file).