Add a script to fix inconsistent references to people from SayIt speakers
We have changed the form of IDs in the South African PopIt instance
several times, but without removing the entries corresponding to old
PopIt IDs from the popit-django popit_person table. This meant that the
name resolution on importing new committee proceedings could associate
the speeches with various SayIt speakers corresponding to different
entries in that table.

This commit introduces a script to hopefully clean up that mess; it
tries to remove any row with an deprecated format ID from the
popit_person table, safely removing any SayIt Speaker or EntityName
associated with that row. It also removes any rows from this table that
can't be associated with any Pombola core Person any more, so long as
they are not referenced any more.

This should fix #1597 and hopefully #1601.

The underlying issue (that Pombola ZA stores the identity of people in
five different places) is being addressed by #1594.