ZA: Reduce queries used on the organisation detail page
Despite the careful use of select_related and prefetch_related in
pombola.core.views:OrganisationDetailView, the loading of an
organization with even a small number of people in it generate a huge
number of queries.

This is caused by the additional information requested by the South
African override of core/person_list_item.html, which does two costly
operations:

 1. Finding places associated with that person by iterating over
    position_set.all.current_unique_places and then accessing
    parent_place for each of those too.

 2. Calling the 'parties' method on each person, which creates a new
    queryset for each person.

Neither of these are all that easy to make more efficient while still
providing the same information without restructuring a lot of the
template code, so instead this commit just paginates the results so that
the page load times don't become excessive. (This reduces the number of
queries for a page load from thousands to about 80; still much higher
than we'd like, but the performance is OK.)

Fixes #1610