Reduce the queries needed for the MPs page
'positions' in the context has been created with appropriate
select_related / prefetch_related invocations to reduce the number of
queries needed to render the MPs page - this commit switches the
template to use this rather than going back to object.position_set to
find the positions. This speeds up the load time of this page.