Add basic word cloud implementation to main pombola directory.
The cloud is based on the last 20 Entry objects. There were a couple
of attempts at this in the code already, but
it hadn't been working and was effectively overridden by code
which just used the search index of all documents.

To do this I've added sitting_start_date to the search index for
hansard entries. This index will need rebuilding with

python manage.py rebuild_index