Update content of expected test files.
This is mostly changing them to be UTF-8 encoded rather than having
reprs of strings in them - since UTF-8 is the default encoding for JSON
files this should all be fine.

I've also stripped the expected output in one test so that it behaves
properly if a newline appears at the end of the file.