Allow caching of the wordcloud via a management command.
Add a couple of settings to store the location of a wordcloud
cache file in a subdirectory of the root_dir.

The wordcloud now uses the file at wordcloud_cache/wordcloud.json
if it exists, and dynamically creates it otherwise.

The management command creates the directory if it is missing and
then populates it with a wordcloud file.