Update za-hansard to a version with PopIt ID blacklisting
We've updated za-hansard to blacklist IDs that are
problematic in terms of misattribution, so we can reenable the committee
scraper.

This also adds --delete-existing to the committee scraper, since one
problematic committee meeting will need to be reimported from the JSON;
if that works, it can be removed tomorrow.

(Pair-programmed with Duncan Parkes.)