KE: Add time-on-page tracking for the Youth Employment page
This uses riveted.js to send a ping to the server every 5 seconds
when it appears that the user is on the page.  riveted is quite
smart about this, looking at whether people click, scroll, etc.
I'm largely using this because Nick has elsewhere on our sites,
so I guess it meets with his approval.

This is only enabled for the Youth Employment Bill experiment page
at the moment.