Update the embedded YouTube clips to edited ones
Since we can't upload video to the Mzalendo YouTube account these
are links to unlisted videos on the mySociety YouTube account.

Thanks to Dave Whiteland (@davewhiteland) for doing the video
editing to cut down the longer videos to these two clips.

This is a temporary measure; we should use versions of these videos
hosted by the Mzalendo YouTube account instead.