KE: Youth Employment experiment: fix GA's first redirection to variant 'y'
The Google Analytics experiment was exhibiting odd behaviour: when
it picked the 'y' variant, the page it would reload to showed the 'n'
variant instead, but on subsequent reloads it would correctly show
the 'y' variant.

This was due to the context being updated from session before the
"first page load for this user with utm_expid' in the GET parameters
case".  This could probably alternatively be fixed by moving that case
above the: "context.update(self.get_session_data())" line, but in the
circumstances this seems more conservative.