Make the initial size of the survey window slightly larger vertically
As requested by MIT; the size of the survey window happened to open up
on many browsers with default text size such that you would need to
scroll to see the rest of the page, but the bottom of the window fell
exactly between paragraphs so it was easy to miss that you had to
scroll.  This commit makes the suggested size of the opened window
slightly larger so that the end of the window falls in the middle of a
line of text, on Chrome anyway.