Stop using Google Analytics' Content Experiments for randomizing pages
All the data that MIT want is being stored in our database, and we're
not using Google Analytics for anything except picking a random variant
of the page.  Using it for randomization also has a few difficulties:

 * It does a redirect after the page load, and sometimes we don't seem
   to be getting that redirect (because of Javascript failures or
   network problems, we assume)

 * There's not a natural goal we can set for the content experiment, so
   we've just been setting 99.5% confidence that one variant increases
   page views across the site as a whole, which is unlikely to happen
   and stop the experiment, but still seems silly.

 * The code in the view is unnecessarily complicated to deal with
   essentially ignoring page views without a utm_expid.

So, this commit switches to picking a variant randomly ourselves, so
there should only be a single page load after clicking the Facebook
advertisement.