Youth Employment experiment: include submitted comments in CSV
This is a quick hack to include the submitted comments on the Youth
Employment page in the exported CSV; the hackiness is due to the
original data modelling not being good (see the comments in this
commit). This will allow that data to be included which is urgently
needed by MIT / Mzalendo, but in the future this data should be
migrated from feedback to the experiment's events.