NG: Fix the PU number lookup
This was breaking because the regex that was checking for a valid
PU number was expecting a PU formatted string and we were passing
it the raw query instead, so the PU lookup was being bypassed
unless the search string took the form of OD:1:2.

Although we don't do anything with the last part of a full 4 part
PU number, we are now accepting the full 01:02:03:04 format as
input for our PU number lookup.