Rename PopoloOtherName to PopoloPersonOtherName
The django-popolo OtherName class is a generic relation, which can be
used for alternative names for organizations, say, as well as persons.
We're only using it for alternative names for people, though, and to
avoid losing data we added the extended person-specific Popolo name
fields - the name of this model should reflect that now it's specific
to a person by convention.

Pair-programmed with Duncan Parkes (@duncanparkes)