ZA: Add Facebook comment support
Adds ZA_FACEBOOK_APP_ID as a config setting then uses that in
preference to any Disqus config values to pull in the Facebook
comments plugin instead.

The Facebook equivalent of the comment thread identifier has to
be a fully formed URL and is currently set using
request.build_absolute_uri. However, the url doesn't have to match
that of the page it is on, which could be useful if we wanted to
include the same discussion on multiple pages.

A word of caution - changing which site the comments are associated
with in the Facebook app (and the domains have to match or it won't
work) seems to lead to comment loss.