Include place data for positions in the Popolo JSON output
Some positions (e.g. those representing an MP for a particular
constituency) have a place attribute. This wasn't being exported
in the nightly JSON export, however. This commit introduces a change
so that the place data is exported as the 'area' attribute of a
Popolo membership.  In addition, if the place is specific to a
particular parliamentary session (as happens with constituencies in
Mzalendo) the parliamentary session data is included under the
'session' attribute of the area.

Fixes #1688