Link to the original Disqus discussions
Uses a local file lookup to check the output of the public API (to
avoid seeing unapproved or deleted comments) to check whether to
include the Disqus comments as well as the Facebook ones.

The idea behind the local file is to avoid having to make a live API
call on every page that *could* possibly display the Disqus comments
panel for information we're not expecting to change once we've closed
the Disqus threads to new comments.