Remove an assertion that we've changed our mind about
This assertion said that a deceased person's party membership
shouldn't be listed as a "former" position on their person page.
However, after discussing this with Liz Conlan (@lizconlan), who
introduced this code, we think it's OK to leave this assertion
out now.  (A deceased person is still listed as a member of their
last party in the profile box above - this is just about what
appears in the list of former positions below.)