ZA Display attendance data for members going to committees.
On each person page for ZA, display some stats about that
person's attendance at committees for the years covered
by the PMG API. For now, this is a sentence for each year
showing how many meetings the attended, how many they were
on the invite list for, and the percentage.

For simplicity of display, we're treating all attendances as
the same, so making no distinction between meetings attended
in full and those where the person arrived late and/or departed
early.