Store PMG api itentifiers for people.
To save us having to search for the person page by our own
slug every time, let's save the PMG id of each person and use
that if we have it.

I'm not absolutely certain what's the best thing to store here.
It feels like the identifier is more useful in general, but then
we have to know the format of the attendance url. Perhaps the best
thing would actually be to store both.