ZA: Revert person_detail.html back to something more like the default
Notably, contact details other than the email and twitter handle
are now back in the sidebar. This shows the most significant
organisations that they have a current position in right under
their name.

(There are only three organizations in important_organizations and
I don't think someone can be both a member of the NCOP and National
Assembly, so there should be a maximum of two of these "significant
organisations", meaning we don't need to worry about "comma and and"
logic - we just join the organization names with ' and '.)

This commit includes removing Peoples Assembly’s person page override
styles.

The position display was contributed by Mark Longair (@mhl)