Don't use ZA views at testing time unless testing for ZA.
At testing time, something very odd was happening - even when using
settings which didn't have South Africa as the country, the South
Africa Specific organisation and person view classes were being used
(I spotted this because the unit tests were trying to download from
the live PMG API, but for all countries, not just ZA).

The problem turned out to be that the organisation and person urls
from core that South Africa wanted to change the view on were being
modified in place at import time on the pombola/south_africa/urls.py
module. When running the site I think this wasn't getting imported
for COUNTRY_APPs other than South Africa, but when testing, nose
imports all the modules so these changes were being made despite
the country app.

Instead of modifying the patterns in place, we now import the
person patterns, organisation patterns and place patterns,
take copies of them, modify these, and then include them
at the same URL prefix as before. This preserves the order
of the URLs in each of these pattern lists and therefore avoids
the problem of the '/(slug)' patterns getting triggered before the
'/all' patterns (see caa8d1cb16f76d0b and 259a4377b19f1140).

I've also added a '$' to the end of the person sub-page urls
to stop them matching more specific things with stuff after the
final slash.

The reduction in the number of queries in one of the tests is a
symptom of this - extra South Africa specific context was being
added even when the country was not South Africa, resulting in
three additional queries.