Avoid FOUC on ZA person page tabs
Rather than duplicating styles (which just gives the CSS engine
more cruft to deal with) we just apply the classes that jQuery
UI will add, at pageload. Meaning the tabs are styled, but not
activated.

To keep everything accessible when JavaScript is disabled, all
of the tab panels are visible, until hidden by jQuery UI. This
still results in a flash of all four tabs before the last three
are hidden, but it’s worth it to maintain accessibility.