Fix equality check for summary fields.
The summary field on Person objects was failing the equality check
for merging people even when it was equal in all the important ways.
Just comparing the .summary on both objects is not good enough
because it's a complex field that holds a reference to the object
which it is a summary of. Instead we have to compare the .rendered
attributes of those summary fields.