Fix equality check for summary fields.
The summary field on Person objects was failing the equality check
for merging people even when it was equal in all the important ways.
Just comparing the .summary on both objects is not good enough
because it's a MarkupField. Instead we have to compare the .rendered
attributes of those summary fields.