The scheme parameter is always empty now
All invocations (from views) of the pombola_person_to_sayit_speaker method
just passed in the empty string as the scheme parameter, so we can
assume that's always the case.  This also means that we can remove the
--sayit-id-scheme option from core_merge_people, and don't need to try
to remove and Identifier with that scheme since the Identifier model's
scheme field isn't allowed to be blank.