Choose important organisations based on kind.
Rather than having a list of org slugs which are considered
important, have a list of org kind slugs which are important.

This gets us the provincial legislatures, and also allows for,
should it ever happen, the addition of something like another house.

A further improvement to this would be hold what organisation
kinds are this sort of important in a database table which
can be changed from the Django admin, getting this into data
and out of code.

Once we have some tests, this fixes #1857.