Create a public pg_dump of "safe" (i.e. no user data) tables every day
It's useful for getting a developer up-and-running with Pombola to
have a database dump available; it would also be useful as part of
the provisioning of a Vagrant box to be able to restore a database
dump to get a realistic site.  This commit gets cron to do a pg_dump
of the database daily which should be useful in both of these
situations.

The risk here is that there is personal information in the database
(e.g. the auth_user table, the django_session table, feedback
submitted to the site, etc.)  The script that performs the database
dump includes a list of tables to exclude (e.g. those containing
personal data) and a list of tables that are expected to be included
in the dump.  If there are any newly added tables (i.e. those that
have been neither marked as "expected" or "to ignore" the script
exits in error so that we can mark that table appropriately.

This initial version runs them between 3pm and 5pm, so that we can check
the output in working hours - we'll switch it back after that testing.

Fixes #1891