NG: Recreate the senatorial districts, but correctly
The boundaries for senatorial districts currently in the ShineYourEye
database were formed from composing the LGA boundaries.  Unfortunately,
this seems to have been done incorrectly since there were lots of
obvious errors in the boundaries.  This script adds synonyms for various
aliases to make name matching work correctly and then recreates each
senatorial district's geometry.

It also changes the parent of each LGA to be the senatorial district,
and the parent of each senatorial district to be the state - this
inclusion is a strict hierarchy, so this is more useful than previously
where the parent of an LGA was a state.