Include ward councillors on the rep locator page
This adds ward councillors to the rep locator page, getting the data
from the Code4SA API, and rewrites the view code to incorporate that
and hopefully make it easier to understand.

The people are divided into two sections: "Ward Councillors" and
"MPs and MPLs".

In addition, this switches the page to be person-centric rather than
office-centric.