NG: Add a command for fixing Federal Constituency boundaries and names
There are instructions for using this in the README.md, updated in
this commit. It depends on the results of the previous scripts
described there.

A couple of the boundaries are not correct; these are cases where
the Federal Constituency (FED) boundary cannot be composed from LGA
boundaries.  These instances are:

  Lagos Island (LGA)
      which covers both LAGOS ISLAND I (FED) and LAGOS ISLAND II (FED)
  Mushin (LGA)
      which covers both MUSHIN I (FED) and MUSHINI II (FED)
  Oshodi/Isolo (LGA)
      which covers both OSHODI/ISOLO I (FED) and OSHODI/ISOLO II (FED)
  Surelere (LGA) (the one in Lagos Central)
      which covers SURELERE I (FED) and SURULERE II (FED)
  Port Harcourt (LGA)
      which covers PORT HARCOURT I (FED) and PORT HARCOURT II (FED)

In these cases, we've used the LGA boundary for both FED boundaries,
so they're too large and completely overlap.  Services that deal with
MapIt results for these areas will get both FED results returned for
point-in-polygon lookups, and should warn about this.