Avoid errors viewing SayIt pages when logged in as a superuser
We were getting errors when someone's logged in who's marked as
a superuser because the template tries to show "edit" and "delete"
links for speeches in those cases.

Since we have no plans at the moment to allow people to edit or
delete speeches on the site, just work around this by forcing
is_user_instance to always be False.