Don't trust sys.path[0] to find the relative location of data files
These management commands assumed that sys.path[0] would be the root
of the Django project, but it doesn't have to be (e.g. if you've done
something odd with pip it might be a 'src' subdirectory of the project
root).  It's clearly more reliable to base the data directory's
location on __file__ anyway.