Switch to container-based builds on Travis
The "sudo: false" means that Travis will try to run to run its test on
their new container-based infrastructure. This means that builds will
be much faster to start up, which will hopefully remove a lot of
frustration with waiting for tests to run on the old infrastructure.

Note that this removes the Python GDAL requirement, which it turns out
is not required anyway and creates problems on container-based builds
on Travis. [1]  (It was only there for MapIt, but it's not required
by MapIt either - Django's code calls the C library directly.)  Thanks
to Matthew Somerville (@dracos) for pointing this out.  Futher helpful
background from him and @jennyd on the complicated situation with the
Python GDAL bindings can be found here:

  https://github.com/alphagov/mapit/commit/2908cda3

[1] https://github.com/travis-ci/travis-ci/issues/2401