Models can define fields to be whitespace stripped
This adds a 'fields_to_whitespace_normalize' property to models which
can be set to a list of fields which should have their whitespace
normalized before the model is saved.

Normalized in this context means removing all of the leading and
trailing whitespace as well as reducing multiple spaces between words to
a single space.